# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart AI Prediksi Togel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results/grouped:
    get:
      operationId: getGroupedResults
      tags: [results]
      summary: Get results grouped by day
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Grouped results by day
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupedDayResult"
  /results:
    get:
      operationId: getResults
      tags: [results]
      summary: Get flat list of lottery results
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of lottery results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LotteryResult"
  /results/stats:
    get:
      operationId: getResultStats
      tags: [results]
      summary: Get lottery statistics
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Statistics for lottery results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultStats"
...
  /smart-ai/analyze:
    get:
      operationId: getSmartAI
      tags: [smart-ai]
      summary: Run Smart AI 7-engine analysis
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
        - in: query
          name: slot
          schema:
            type: string
      responses:
        "200":
          description: Smart AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartAIResult"
  /smart-ai/slots:
    get:
      operationId: getSmartAISlots
      tags: [smart-ai]
      summary: Get all slot analyses for a pasaran
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
      responses:
        "200":
          description: All slot analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartAISlotsResponse"
  /smart-ai/weights:
    get:
      operationId: getSmartAIWeights
      tags: [smart-ai]
      summary: Get Smart AI adaptive engine weights
      parameters:
        - in: query
          name: pasaran
          schema:
            type: string
      responses:
        "200":
          description: Smart AI weights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartAIWeightsResponse"
components:
  schemas:
...
    LotteryResult:
      type: object
      required: [id, pasaran, tanggal, periode, result4d, as, kop, kepala, ekor, result2d, result3d]
      properties:
        id:
          type: integer
        pasaran:
          type: string
        tanggal:
          type: string
        periode:
          type: string
        result4d:
          type: string
        as:
          type: string
        kop:
          type: string
        kepala:
          type: string
        ekor:
          type: string
        result2d:
          type: string
        result3d:
          type: string
...
    SlotMap:
      type: object
      properties:
        "00:01":
          type: ["string", "null"]
        "13:00":
          type: ["string", "null"]
        "16:00":
          type: ["string", "null"]
        "19:00":
          type: ["string", "null"]
        "22:00":
          type: ["string", "null"]
        "23:00":
          type: ["string", "null"]
    GroupedDayResult:
      type: object
      required: [tanggal, hari, slots]
      properties:
        tanggal:
          type: string
        hari:
          type: string
        slots:
          $ref: "#/components/schemas/SlotMap"
    ResultStats:
      type: object
      required: [pasaran, totalDraws, digitFrequency, positionFrequency, hotNumbers, coldNumbers, lastUpdated]
      properties:
        pasaran:
          type: string
        totalDraws:
          type: integer
        digitFrequency:
          type: object
          additionalProperties:
            type: integer
        positionFrequency:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: integer
        hotNumbers:
          type: array
          items:
            type: string
        coldNumbers:
          type: array
          items:
            type: string
        lastUpdated:
          type: string
        oddEvenDist:
          type: object
          properties:
            odd:
              type: integer
            even:
              type: integer
        bigSmallDist:
          type: object
          properties:
            big:
              type: integer
            small:
              type: integer
        twoDigitFrequency:
          type: object
          additionalProperties:
            type: integer
...
    SmartEngineOutput:
      type: object
      required: [id, name, weight, top5, confidence, scores2d]
      properties:
        id:
          type: string
        name:
          type: string
        weight:
          type: number
        top5:
          type: array
          items:
            type: string
        confidence:
          type: number
        scores2d:
          type: object
          additionalProperties:
            type: number
    SmartColokJitu:
      type: object
      required: [posisi, digit]
      properties:
        posisi:
          type: string
        digit:
          type: string
    SmartRekomend:
      type: object
      required: [rank, number, score, label, color]
      properties:
        rank:
          type: integer
        number:
          type: string
        score:
          type: number
        label:
          type: string
        color:
          type: string
    SmartAIResult:
      type: object
      required: [pasaran, slot, analyzedAt, totalSlotDraws, lastResult, lastResultAt, overallConfidence, main4d]
      properties:
        pasaran:
          type: string
        slot:
          type: string
        analyzedAt:
          type: string
        totalSlotDraws:
          type: integer
        lastResult:
          type: string
        lastResultAt:
          type: string
        prevSlotResult:
          type: string
        overallConfidence:
          type: number
        main4d:
          type: string
        bbfs5d:
          type: array
          items:
            type: string
        pred4d:
          type: array
          items:
            type: string
        pred3d:
          type: array
          items:
            type: string
        pred2dEkor:
          type: array
          items:
            type: string
        pred2dDepan:
          type: array
          items:
            type: string
        pred2dTengah:
          type: array
          items:
            type: string
        colokBebas:
          type: array
          items:
            type: string
        colokBebas2d:
          type: array
          items:
            type: string
        colokJitu:
          type: array
          items:
            $ref: "#/components/schemas/SmartColokJitu"
        dasar:
          type: string
        tengahTepi:
          type: string
        silangHomo:
          type: string
        kembangKempis:
          type: string
        shio:
          type: array
          items:
            type: string
        bom2d:
          type: string
        topDigitKandidat:
          type: array
          items:
            type: string
        topRekomendasi:
          type: array
          items:
            $ref: "#/components/schemas/SmartRekomend"
        engines:
          type: array
          items:
            $ref: "#/components/schemas/SmartEngineOutput"
    SlotCountItem:
      type: object
      required: [slot, count]
      properties:
        slot:
          type: string
        count:
          type: integer
    SmartAISlotsResponse:
      type: object
      required: [pasaran, slots, targetSlot]
      properties:
        pasaran:
          type: string
        targetSlot:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/SlotCountItem"
    EngineWeightInfo:
      type: object
      required: [id, name, defaultWeight, adaptiveWeight, hitRate, evalCount]
      properties:
        id:
          type: string
        name:
          type: string
        defaultWeight:
          type: number
        adaptiveWeight:
          type: number
        hitRate:
          type: number
        evalCount:
          type: integer
    SmartAIWeightsResponse:
      type: object
      required: [pasaran, weights, hitRates, isAdaptive, detail]
      properties:
        pasaran:
          type: string
        weights:
          type: object
          additionalProperties:
            type: number
        hitRates:
          type: object
          additionalProperties:
            type: number
        evalCounts:
          type: object
          additionalProperties:
            type: integer
        isAdaptive:
          type: boolean
        lastUpdated:
          type: ["string", "null"]
        detail:
          type: array
          items:
            $ref: "#/components/schemas/EngineWeightInfo"