# lib/api-zod/src/generated/types/smartRekomend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart AI Prediksi Togel API
 * OpenAPI spec version: 0.1.0
 */

export interface SmartRekomend {
  rank: number;
  number: string;
  score: number;
  label: string;
  color: string;
}